requirements.txt
fastapi
uvicorn
httpx
pydantic-ai-slim[google]
python-dotenv
